+++ packages/fluentui/react-bindings/src/FocusZone/FocusTrapZone.types.tsx
@@ -73,7 +73,7 @@ export interface FocusTrapZoneProps extends React.HTMLAttributes<HTMLDivElement>
    * When FocusTrapZone is released, FocusTrapZone will try to restore focus back to the previously focused element;
    * sometimes focusing on the previously focused element can cause page to scroll. Set this prop to true to prevent scrolling
    */
-  preventScrollOnRestoreFocusToPreviouslyFocusedElement?: boolean;
+  preventScrollOnRestoreFocus?: boolean;
 
   innerRef?: any;
 }